refactor(pagination): :recycle: Add VDPaginator component and refactor ReportsView
